requirements.txt
# Core API Framework
fastapi>=0.95.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6

# AI/ML Libraries
sentence-transformers>=2.2.2
faiss-cpu>=1.7.0
numpy>=1.21.0,<2.0.0  # Will use pre-built wheel
torch>=2.0.0
rank-bm25>=0.2.2  # For hybrid search

# HTTP & Async
httpx>=0.24.0
tenacity>=8.2.0

# Logging & Environment
loguru>=0.7.0
python-dotenv>=0.21.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfminer.six>=20221105

# UI Framework
streamlit>=1.24.0
plotly>=5.14.0
pandas>=1.5.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0

# Database
# SQLite is built-in to Python, no need to install
